feat: Add encoding and compression functionality
This commit adds encoding and compression functionality to the codebase. It introduces a new `Compressor` enum and its associated methods `compress` and `encode`. The `encode` function is implemented for the `Encodable` trait and takes in an input and size parameter to encode the data. The `compress` function takes in a vector of bytes and a compressor enum and compresses the data using the selected compressor. The code changes also include a new `RleEncoder` struct in the `rle.rs` file, which implements the RLE encoding algorithm. The `encode` function is implemented for the `RleEncoder` struct to perform RLE encoding on the data. Additionally, new functions `write_bytes_to_file` and `write_encoded_file` are added to the `io` module for writing bytes to a file and writing an encoded file, respectively. Finally, in the `main.rs` file, the encoding and compression functions are used to encode the quadtree data, compress it using RLE, and write the compressed data to a file.